requirements.txt
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
beautifulsoup4==4.13.4
bidict==0.23.1
blinker==1.8.2
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.2
click==8.1.7
colorama==0.4.6
contourpy==1.3.0
cycler==0.12.1
dnspython==2.7.0
et-xmlfile==1.1.0
fastapi==0.116.1
Flask==3.0.3
Flask-Cors==5.0.0
Flask-SocketIO==5.4.1
fonttools==4.53.1
h11==0.14.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
joblib==1.4.2
kiwisolver==1.4.7
MarkupSafe==2.1.5
matplotlib==3.9.2
numpy==2.1.1
openpyxl==3.1.5
outcome==1.3.0.post0
packaging==24.1
pandas==2.2.2
pillow==10.4.0
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
pydub==0.25.1
pymongo==4.13.2
PyMuPDF==1.26.3
pyparsing==3.1.4
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-engineio==4.10.0
python-socketio==5.11.4
pytz==2024.2
requests==2.32.4
scikit-learn==1.5.1
scipy==1.14.1
selenium==4.34.2
setuptools==80.9.0
simple-websocket==1.1.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.7
starlette==0.47.2
threadpoolctl==3.5.0
tqdm==4.67.0
trio==0.30.0
trio-websocket==0.12.2
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2024.1
urllib3==2.5.0
uvicorn==0.35.0
webdriver-manager==4.0.2
websocket-client==1.8.0
Werkzeug==3.0.4
wsproto==1.2.0
xgboost==3.0.2
